增加ide提示包，和错误码生成 以及多语言出来

+++ app/admin/service/Menu.php
@@ -2,14 +2,23 @@
 
 namespace App\admin\service;
 
+use app\constant\Error;
+use app\exception\AdminException;
 use app\model\AdminPermission;
 use app\model\AdminRole;
 use app\model\AdminUser;
 use App\utils\ParseDocument;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
+use function Symfony\Component\String\s;
 
 class Menu {
+
+    /** @var int 默认排序 */
+    const DEFAULT_ORDER = 255;
+
+    /** @var array 标识集合 */
+    static $slug = [];
 
     /**
      * 获取用户菜单列表
@@ -17,7 +26,7 @@ class Menu {
      * @return array|mixed
      */
      public static function getUserMenus(int $adminId = AdminUser::DEFAULT_ADMIN_ID) {
-         $menus = AdminPermission::allNodes()->toArray();
+         $menus = AdminPermission::menuNodes();
          $menus = self::formatMenus($menus); 
          if ($adminId == AdminUser::DEFAULT_ADMIN_ID) {
              return $menus;
@@ -46,54 +55,30 @@ class Menu {
      }
 
      /**
-      * 菜单最多二级
+      * 格式化菜单
       * @param $menus
       * @return mixed
       */
      private static function formatMenus($menus)
      {
-          foreach ($menus as &$menu) {
-              $menu['meta'] = [
-                  'icon' => $menu['icon'],
-                  'requiresAuth' => true,
-                  'locale' => $menu['locale'],
-              ];
-              $menu['name'] = $menu['slug'];
+         $result = [];
+         foreach ($menus as $menu) {
+             $temp = [
+                 'name' => $menu['slug'],
+                 'meta' => [
+                     'icon' => $menu['icon'],
+                     'requiresAuth' => true,
+                     'locale' => $menu['locale'],
+                 ],
+             ];
 
-              foreach ($menu['children'] as &$children) {
-                   $children['meta'] = [
-                        'icon' => $menu['icon'],
-                        'requiresAuth' => true,
-                        'locale' => $children['locale'],
-                    ];
+             if (count($menu->children) > 0) {
+                 $temp['children'] = static::formatMenus($menu->children);
+             }
 
-                    $children['name'] = $children['slug'];
-                    unset($children['permission_id']);
-                    unset($children['order']);
-                    unset($children['parent_id']);
-                    unset($children['is_menu']);
-                    unset($children['icon']);
-                    unset($children['locale']);
-                    unset($children['slug']);
-                    unset($children['children']);
-                    unset($children['update_time']);
-                    unset($children['delete_time']);
-                    unset($children['create_time']);
-              }
-
-              unset($menu['permission_id']);
-              unset($menu['order']);
-              unset($menu['parent_id']);
-              unset($menu['is_menu']);
-              unset($menu['locale']);
-              unset($menu['icon']);
-              unset($menu['slug']);
-              unset($menu['update_time']);
-              unset($menu['delete_time']);
-              unset($menu['create_time']);
-          }
-
-          return $menus;
+             $result[] = $temp;
+         }
+         return $result;
      }
 
     /**
@@ -119,7 +104,7 @@ class Menu {
      * 初始化后台系统菜单
      * @return mixed
      */
-     public static function init()
+     public static function init($isClear = false)
      {
          $doMenuAction = new ParseDocument([app_path() .DIRECTORY_SEPARATOR . 'admin/controller']);
          $result = $doMenuAction->setStdout(true)->lookup()->getResult();
@@ -139,14 +124,15 @@ class Menu {
          }
 
          static::$slug = [];
-         AdminPermission::truncate();
+         if ($isClear) {
+             AdminPermission::truncate();
+         }
+
          static::saveMenuTreeToDatabase($result);
          static::$slug = [];
 
          return $result;
      }
-
-     static $slug = [];
 
     /**
      * 对遍历菜单进行入库操作
@@ -158,19 +144,10 @@ class Menu {
              if (in_array($parent['slug'], static::$slug)) {
                  continue;
              }
+             $menu = static::addMenu(array_merge($parent, ['parentId' => $parentId]));
 
-             $menu = new AdminPermission([
-                 'name' => $parent['name'],
-                 'slug' => $parent['slug'],
-                 'locale' => $parent['locale'],
-                 'icon' => $parent['icon'],
-                 'path' => $parent['path'],
-                 'order' => $parent['order'] ?? 255,
-                 'parent_id' => $parentId,
-                 'is_menu' => $parent['isMenu'] ? 1 : 0
-             ]);
              static::$slug[] = $parent['slug'];
-             if ($menu->save() && isset($parent['children'])) {
+             if (isset($parent['children'])) {
                  static::saveMenuTreeToDatabase($parent['children'], $menu->permission_id);
              }
          }
@@ -189,4 +166,131 @@ class Menu {
              }
          }
      }
+
+    /**
+     * 获取所有菜单
+     * @return mixed
+     */
+    public static function getMenus() {
+        return AdminPermission::allNodes();
+    }
+
+    /**
+     * 添加菜单或权限
+     * @param $params
+     * @return AdminPermission|false
+     * @throws AdminException
+     */
+    public static function addMenu($params) {
+        $isExist = AdminPermission::where('slug', $params['slug'])->first();
+        if ($isExist) {
+            throw new AdminException(Error::MenuSlugUniqiued);
+        }
+
+        $menu = new AdminPermission([
+            'name' => $params['name'],
+            'slug' => $params['slug'],
+            'locale' => $params['locale'],
+            'icon' => $params['icon'],
+            'path' => $params['path'],
+            'order' => $params['order'] ?: self::DEFAULT_ORDER,
+            'parent_id' => $params['parentId'] ?: 0,
+            'is_menu' => $params['isMenu'] ? AdminPermission::MENU_YES : AdminPermission::MENU_NO
+        ]);
+
+        if ($menu->save()) {
+            return $menu;
+        }
+
+        throw new AdminException(Error::SaveFailed);
+
+    }
+
+    /**
+     * 更新菜单
+     * @param $menuId
+     * @param $data
+     * @return AdminPermission|AdminPermission[]|false|Collection|Model
+     */
+    public static function updateMenu($menuId, $data) {
+        $menu = AdminPermission::find($menuId);
+        if (!$menu) {
+            return false;
+        }
+
+        $menu->name = $data['name'];
+        $menu->icon = $data['icon'];
+        $menu->locale = $data['locale'];
+        $menu->path = $data['path'];
+        $menu->order = $data['order'] ?? $menu->order;
+        $menu->parent_id = $data['parentId'] ?? $menu->order;
+        $menu->is_menu = $data['isMenu'] ? AdminPermission::MENU_YES : AdminPermission::MENU_NO;
+        $menu->slug = $data['slug'] ?? $menu->slug;
+
+        $menu->save();
+
+        return $menu;
+
+    }
+
+    /**
+     * 格式化菜单
+     * @param AdminPermission $permission
+     * @return array
+     */
+    public static function formatBaseMenu(AdminPermission $permission): array
+    {
+        $result = [
+            'permissionId' => $permission->permission_id,
+            'name' => $permission->name,
+            'icon' => $permission->icon,
+            'slug' => $permission->slug,
+            'locale' => $permission->locale,
+            'path' => $permission->path,
+            'order' => $permission->order,
+            'parentId' => $permission->parent_id,
+            'isMenu' => $permission->is_menu
+        ];
+
+        $allChildren = $permission->allChildren;
+        if (count($allChildren) > 0) {
+            foreach ($allChildren as $child) {
+                $result['children'][] = static::formatBaseMenu($child);
+            }
+        }
+        return $result;
+    }
+
+    /**
+     * 格式化菜单实体
+     * @param $list
+     * @return array
+     */
+    public static function formatMenusEntity($list): array
+    {
+        $result = [];
+        foreach ($list as $menu) {
+            $result[] = static::formatBaseMenu($menu);
+        }
+
+        return $result;
+    }
+
+    /**
+     * 删除菜单/权限
+     * @param $menuIds
+     * @return false
+     */
+    public static function deleteMenu($menuIds): bool
+    {
+        $menuIds = array_unique(array_filter($menuIds));
+
+        if (!$menuIds) {
+            return false;
+        }
+
+        return (bool)AdminPermission::query()
+            ->whereIn('permission_id', $menuIds)
+            ->delete();
+    }
 }
